requirements.txt
tvdatafeed
pandas
pandas-ta
python-telegram-bot==13.15